% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emuR_segmentlistExtensions.R
\name{augment}
\alias{augment}
\title{Add information from one output of an emuR::query command to augment anouther}
\usage{
augment(
  dbHandle,
  segs,
  parentSegs,
  add.metadata = FALSE,
  resultType = "tibble"
)
}
\arguments{
\item{dbHandle}{An emuR databse handle.}

\item{segs}{The original \code{\link{emuRsegs}} object.}

\item{parentSegs}{The \code{\link{emuRsegs}} object containg the additional information that should be added to the \code{segs} argument object.}

\item{add.metadata}{Boolean: Should metadata associated with the bundle from which \code{segs} was returned be attached to the output "tibble" as columns? #TODO}

\item{resultType}{Can only be "tibble" at the moment, and the function will therefore return a \code{\link{ti}}.}
}
\value{
A "tibble" that has the columns of the \code{segs}, but with the columns of
\code{parentSegs} appended to the right, where a parent is match by its decendant by
the part of the time line of the signal that they share. A prefix "parent_" is added
to all the columns in \code{parentSegs}. That is, a label A will be on the same row
as the parent P if, for instance, sample_start > parent_sample_start and
sample_end < parent_sampe_end in the output.
}
\description{
This function may be used to attach additional information to an \code{\link{emuRsegs}}
object. The additional information comes from a second \code{\link{emuRsegs}}
containing annotations that are ancestors of labels in the first \code{\link{emuRsegs}} object. That is, the ancestors should span the decendent labels in time. The labels
that add information could come for instance from a \code{\link{requery_hier}} call
but could also be the result of an entierly separate query.
}
\details{
It is important to note that the information in the two \code{\link{emuRsegs}}
objects are merged by start and stop time, and not by hierarchy.
}
\examples{
create_ae_db() -> ae
query(ae,"Word =~ .*") -> wd
query(ae,"Utterance =~ .*") -> wd
augment.emuRsegs(ae,wd,utt) \%>\% glimpse()
unlink_emuRDemoDir()
rm(ae)

}
