% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rythm.R
\name{COV}
\alias{COV}
\title{A simple utility function to compute the coefficient of variance for values in a vector}
\usage{
COV(x, na.rm = TRUE)
}
\arguments{
\item{x}{The vector to process.}

\item{na.rm}{boolean; Should NAs be removed for the computations. Otherwise, an NA will be returned for a vector containing one or more NAs.}
}
\value{
The standard deviation / mean of a vector
}
\description{
This function now uses a high-performance Rcpp implementation for improved speed.
}
\author{
Fredrik Karlsson
}
