% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rythm.R
\name{COV5_x}
\alias{COV5_x}
\title{Computes the coefficient of variance for an interval vector based on the average of the first four elements in the vector of intervals.}
\usage{
COV5_x(x, n = 20, return.na = TRUE, na.rm = TRUE)
}
\arguments{
\item{x}{The input vector containing syllable durations.}

\item{n}{The number of intervals to include in the computations. Defaults to 20. The vector should not contain missing values.}

\item{return.na}{boolean;Return NA in the case there are not 8 intervals in the series?}

\item{na.rm}{Should NA intervals be removed?}
}
\value{
A single value indicating Relative coefficient of variation in intervals $5..n$ relative to the mean duration of intervals $1..4$.
}
\description{
Computes the coefficient of variance for an interval vector based on the average of the first four elements in the vector of intervals.
}
\references{
Skodda, S., Lorenz, J., & Schlegel, U. (2012). Instability of syllable repetition in Parkinson's disease—Impairment of automated speech performance? Basal Ganglia, 3(1), 33–37. doi:10.1016/j.baga.2012.11.002
Skodda, S., Schlegel, U., Hoffmann, R., & Saft, C. (2013). Impaired motor speech performance in Huntington’s disease. Journal of Neural Transmission, 1–9–9. doi:10.1007/s00702-013-1115-9
}
\author{
Fredrik Karlsson
}
