% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_cov}
\alias{cpp_cov}
\title{Coefficient of Variation (Rcpp implementation)}
\usage{
cpp_cov(x, na_rm = TRUE)
}
\arguments{
\item{x}{Numeric vector}

\item{na_rm}{Should NA values be removed? Default TRUE}
}
\value{
Coefficient of variation or NA
}
\description{
Computes the coefficient of variation (sd/mean) for a numeric vector.
This is a high-performance replacement for the pure R version.
}
\examples{
x <- c(1.2, 1.5, 1.3, 1.4, 1.6)
cpp_cov(x)
}
\author{
Fredrik Karlsson
}
