% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_cov5_x}
\alias{cpp_cov5_x}
\title{Relative Coefficient of Variation (intervals 5-n vs 1-4)}
\usage{
cpp_cov5_x(x, n = 20L, return_na = TRUE, na_rm = TRUE)
}
\arguments{
\item{x}{Vector of syllable durations}

\item{n}{Number of intervals to include (default 20)}

\item{return_na}{Return NA if insufficient data? (default TRUE)}

\item{na_rm}{Remove NA values before computation? (default TRUE)}
}
\value{
Relative COV or NA
}
\description{
Computes the coefficient of variance for intervals 5-n relative to 
the mean duration of intervals 1-4, as described in Skodda et al. (2012).
}
\examples{
durations <- runif(25, 0.1, 0.3)
cpp_cov5_x(durations)
}
\references{
Skodda, S., Lorenz, J., & Schlegel, U. (2012). Instability of syllable 
repetition in Parkinson's disease. Basal Ganglia, 3(1), 33-37.
}
\author{
Fredrik Karlsson
}
