% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_left_changepoint}
\alias{cpp_left_changepoint}
\title{Find first changepoint in missing value pattern}
\usage{
cpp_left_changepoint(x, what_na = 0)
}
\arguments{
\item{x}{Numeric vector}

\item{what_na}{Threshold for missing values}
}
\value{
Index of first changepoint (1-based) or NA if no changepoint
}
\description{
Identifies the index of the first transition between missing and 
non-missing values (or vice versa).
}
\examples{
x <- c(0, 0, 1.2, 1.5, 2.0)
cpp_left_changepoint(x)  # Returns 3
}
\author{
Fredrik Karlsson
}
