% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_missing_frac}
\alias{cpp_missing_frac}
\title{Compute fraction of missing values}
\usage{
cpp_missing_frac(x, what_na = 0)
}
\arguments{
\item{x}{Numeric vector}

\item{what_na}{Threshold for missing values}
}
\value{
Fraction of missing values (0 to 1)
}
\description{
Calculates the proportion of values that are missing or invalid.
}
\examples{
x <- c(1.2, 0, 1.5, NA, -1, 2.0)
cpp_missing_frac(x)
}
\author{
Fredrik Karlsson
}
