% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_missing_vec_character}
\alias{cpp_missing_vec_character}
\title{Identify missing or invalid values in character vector}
\usage{
cpp_missing_vec_character(x, what_na)
}
\arguments{
\item{x}{Character vector}

\item{what_na}{Character vector of values to consider missing}
}
\value{
Logical vector indicating missing values
}
\description{
Identifies values that are NA or match specified strings.
}
\examples{
x <- c("a", "missing", "b", NA, "c")
cpp_missing_vec_character(x, c("missing", ""))
}
\author{
Fredrik Karlsson
}
