% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_missing_vec_numeric}
\alias{cpp_missing_vec_numeric}
\title{Identify missing or invalid values in numeric vector}
\usage{
cpp_missing_vec_numeric(x, what_na = 0)
}
\arguments{
\item{x}{Numeric vector}

\item{what_na}{Threshold value (values <= this are considered missing)}
}
\value{
Logical vector indicating missing values
}
\description{
Identifies values that are NA, NULL, or below/equal to a threshold.
}
\examples{
x <- c(1.2, 0, 1.5, NA, -1, 2.0)
cpp_missing_vec_numeric(x, what_na = 0)
}
\author{
Fredrik Karlsson
}
