% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_right_changepoint}
\alias{cpp_right_changepoint}
\title{Find last changepoint in missing value pattern}
\usage{
cpp_right_changepoint(x, what_na = 0)
}
\arguments{
\item{x}{Numeric vector}

\item{what_na}{Threshold for missing values}
}
\value{
Index of last changepoint (1-based) or NA if no changepoint
}
\description{
Identifies the index of the last transition between missing and
non-missing values (or vice versa).
}
\examples{
x <- c(0, 0, 1.2, 1.5, 0, 0)
cpp_right_changepoint(x)  # Returns 5
}
\author{
Fredrik Karlsson
}
