% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_vowel_angles}
\alias{cpp_vowel_angles}
\title{Compute vowel vector angles (polar coordinates)}
\usage{
cpp_vowel_angles(f1, f2, f1c, f2c)
}
\arguments{
\item{f1}{F1 formant values (Hz)}

\item{f2}{F2 formant values (Hz)}

\item{f1c}{F1 coordinate of vowel space center}

\item{f2c}{F2 coordinate of vowel space center}
}
\value{
Vector of angles in radians (-pi to pi)
}
\description{
Computes the angle of each vowel vector from the vowel space center,
in radians using atan2.
}
\examples{
f1 <- c(300, 600, 600, 300)
f2 <- c(2200, 1700, 1000, 900)
center <- cpp_vowel_center(f1, f2)
cpp_vowel_angles(f1, f2, center$f1c, center$f2c)
}
\author{
Fredrik Karlsson
}
