% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_vowel_center}
\alias{cpp_vowel_center}
\title{Compute vowel space center}
\usage{
cpp_vowel_center(f1, f2, method = "wcentroid", na_rm = TRUE)
}
\arguments{
\item{f1}{F1 formant values (Hz)}

\item{f2}{F2 formant values (Hz)}

\item{method}{Center computation method: "centroid", "twomeans", or "wcentroid"}

\item{na_rm}{Remove NA values? (default TRUE)}
}
\value{
List with f1c and f2c center coordinates
}
\description{
Computes the center of a vowel space using different methods.
}
\examples{
f1 <- c(300, 600, 600, 300)
f2 <- c(2200, 1700, 1000, 900)
cpp_vowel_center(f1, f2)
}
\author{
Fredrik Karlsson
}
