% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_vowel_norms}
\alias{cpp_vowel_norms}
\title{Compute vowel vector norms (distances from center)}
\usage{
cpp_vowel_norms(f1, f2, f1c, f2c)
}
\arguments{
\item{f1}{F1 formant values (Hz)}

\item{f2}{F2 formant values (Hz)}

\item{f1c}{F1 coordinate of vowel space center}

\item{f2c}{F2 coordinate of vowel space center}
}
\value{
Vector of Euclidean distances
}
\description{
Computes the Euclidean distance of each vowel from the vowel space center.
}
\examples{
f1 <- c(300, 600, 600, 300)
f2 <- c(2200, 1700, 1000, 900)
center <- cpp_vowel_center(f1, f2)
cpp_vowel_norms(f1, f2, center$f1c, center$f2c)
}
\author{
Fredrik Karlsson
}
