% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voice_function_displays.R
\name{hz2st}
\alias{hz2st}
\title{Convert frequency (Hz) to semitones}
\usage{
hz2st(f)
}
\arguments{
\item{f}{Frequency in Hz (numeric vector)}
}
\value{
Semitone values (numeric vector)
}
\description{
Converts frequency values in Hertz to semitones relative to C0 (16.352 Hz).
}
\examples{
hz2st(440)  # A4 = 57 semitones
hz2st(c(220, 440, 880))
}
