% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{jitter_ddp}
\alias{jitter_ddp}
\title{Computes the Difference of Differences of Periods (DDP) of a vector.}
\arguments{
\item{x}{The input vector}

\item{min.period}{The minimum value to be included in the calculation.}

\item{max.period}{The maximum value to be included in the calculation.}

\item{absolute}{Should the Jitter DDP value be returned (absolute=FALSE), or the absolute Jitter DDP (absolute=TRUE). 
In the case of absolute Jitter DDP, the jitter will *not* be devided by the average period.}

\item{omit}{Should missing intervals be removed?}
}
\value{
The Jitter DDP (in s) (absolute=FALSE) or the absolute jitter DDP (in percent, $1..600%$) (absolute=TRUE). If the vector contains less than three values, NA is returned.
}
\description{
Computes the Difference of Differences of Periods (DDP) of a vector.
}
\author{
Fredrik Karlsson
}
