% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{jitter_local}
\alias{jitter_local}
\title{Computes the local jitter of a vector.}
\arguments{
\item{x}{The input vector}

\item{min.period}{The minimum value to be included in the calculation.}

\item{max.period}{The maximum value to be included in the calculation.}

\item{absolute}{Should the (local) Jitter value be returned (absolute=FALSE), or the absolute (local) Jitter (absolute=TRUE). In the case of absolute (local) Jitter, the jitter will *not* be devided by the average period.}

\item{omit}{Should missing intervals be removed?}
}
\value{
A value indicating the (local) jitter (absolute=FALSE) or the absolute (local) jitter (absolute=TRUE).
}
\description{
Computes the local jitter of a vector.
}
\author{
Fredrik Karlsson
}
