% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{jitter_ppq5}
\alias{jitter_ppq5}
\title{Computes the five-point Period Perturbation Quotient (PPQ5) of a vector.}
\arguments{
\item{x}{The input vector}

\item{min.period}{The minimum value to be included in the calculation.}

\item{max.period}{The maximum value to be included in the calculation.
@param absolute Should the Jitter RAP value be returned (absolute=FALSE), or the absolute Jitter PPQ5 (absolute=TRUE). In the case of absolute Jitter PPQ5, the jitter will *not* be devided by the average period.}

\item{omit}{Should missing intervals be removed?}
}
\value{
Either the jitter PPQ5 (in percent, 1..400%) or the absolute PPQ5 values (in s, 1..4) for the vector of values. If the vector contains less than five values, NA is returned.
}
\description{
Computes the five-point Period Perturbation Quotient (PPQ5) of a vector.
}
\author{
Fredrik Karlsson
}
