% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{jitter_rap}
\alias{jitter_rap}
\title{Computes the Relative Average Perturbation (RAP) of a vector.}
\arguments{
\item{x}{The input vector}

\item{min.period}{The minimum value to be included in the calculation.}

\item{max.period}{The maximum value to be included in the calculation.}

\item{absolute}{Should the Jitter RAP value be returned (absolute=FALSE), or the absolute Jitter RAP(absolute=TRUE). In the case of absolute Jitter RAP, the jitter will *not* be devided by the average period.}

\item{omit}{Should missing intervals be removed?}
}
\value{
The RAP (in percent, $1..200%$) or absolute PPQ5 (\code{absAP}) values (in s) for the vector of values. If the vector contains less than three values, NA is returned.
}
\description{
Computes the Relative Average Perturbation (RAP) of a vector.
}
\author{
Fredrik Karlsson
}
