% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nPVI}
\alias{nPVI}
\title{Normalized pairwise variability index.}
\arguments{
\item{x}{A vector of durations in arbitrary unit.}

\item{omit}{Boolean indicating whether NA values should be removed before calculating nPVI.}
}
\value{
A single value reprenting the nPVI for the vector of durations
}
\description{
Computes the normalized Pairwire Variability Index (nPVI) on a supplied vector of durations.
}
\references{
Nolan, F., & Asu, E. L. (2009). The Pairwise Variability Index and Coexisting Rhythms in Language. Phonetica, 66(1-2), 64–77. doi:10.1159/000208931
}
\author{
Fredrik Karlsson
}
