% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rPVI}
\alias{rPVI}
\title{Raw pairwise variability index.}
\usage{
rPVI(x, narm)
}
\arguments{
\item{x}{A vector of durations in arbitrary unit.}

\item{na.rm}{Boolean indicating whether NA values should be removed before calculating rPVI.}
}
\value{
A single value reprenting the rPVI for the vector of durations
}
\description{
Computes the raw Pairwire Variability Index (rPVI) on a supplied vector of durations.
}
\references{
Nolan, F., & Asu, E. L. (2009). The Pairwise Variability Index and Coexisting Rhythms in Language. Phonetica, 66(1-2), 64–77. doi:10.1159/000208931
}
\author{
Fredrik Karlsson
}
