% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfd.R
\name{vowel.angles}
\alias{vowel.angles}
\title{Computes vector angles of vowel vectors of vowels.The vowel space is constructed by the supplied F1 and F2 formant values.}
\usage{
vowel.angles(f1, f2, na.rm = TRUE, center = NULL, center.method = "wcentroid")
}
\arguments{
\item{f1}{A vector of F1 values.}

\item{f2}{A vector of F2 values.}

\item{na.rm}{Boolean value specifying whether missing values should be removed. Defaults to TRUE.}

\item{center}{A vector (F2,F1) specifying the vowel space center to be used. If NULL, the center will be calculated using the \code{\link{vowelspace.center}} function.}

\item{center.method}{The method to use in the calculation of vowel space center. See \code{\link{vowelspace.center}} for details.}
}
\value{
A vector containing the vector angles of each vowel space vector drawn from the origin of teh computed vowel space center coordinates.
}
\description{
Computes vector angles of vowel vectors of vowels.The vowel space is constructed by the supplied F1 and F2 formant values.
}
\examples{
data(pb)
with(pb,vowel.angles(F1,F2))
with(with(pb,data.frame(angle=vowel.angles(F1,F2),vowel=Vowel)),tapply(angle,list(vowel),mean,na.rm=TRUE))

}
\references{
\insertRef{Karlsson:2012vb}{articulated}
}
\seealso{
See \code{\link{vector.space}} for details concerning the computations.
}
\author{
Fredrik Karlsson
}
\keyword{arith}
\keyword{misc}
\keyword{utilities}
