% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfd.R
\name{vowelspace.corners}
\alias{vowelspace.corners}
\title{Computes corners of a vowel space. The vowel space is constructed from the supplied F1 and F2 formant values for vowels in the space.}
\usage{
vowelspace.corners(
  f1,
  f2,
  na.rm = TRUE,
  center = NULL,
  center.method = "wcentroid"
)
}
\arguments{
\item{f1}{A vector of F1 values.}

\item{f2}{A vector of F2 values.}

\item{na.rm}{Boolean value specifying whether missing values should be removed. Defaults to TRUE.}

\item{center}{A vector (F2,F1) specifying the vowel space center to be used. If NULL, the center will be calculated using the \code{\link{vowelspace.center}} function.}

\item{center.method}{The method to use in the calculation of vowel space center. See \code{\link{vowelspace.center}} for details.}
}
\value{
A vector containing the vector length of each vowel space vector drawn from the origin of teh computed vowel space center coordinates.
}
\description{
Computes corners of a vowel space. The vowel space is constructed from the supplied F1 and F2 formant values for vowels in the space.
}
\examples{
data(pb)
with(pb,vowelspace.corners(F1,F2))

}
\references{
Karlsson, F., & van Doorn, J. (2012). Vowel formant dispersion as a measure of articulation proficiency. The Journal of the Acoustical Society of America, 132(4), 2633–2641. doi:10.1121/1.4746025
}
\seealso{
See \code{\link{vector.space}} for details concerning the computations.
}
\author{
Fredrik Karlsson
}
\keyword{arith}
\keyword{misc}
\keyword{utilities}
