% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfd.R
\name{VSA}
\alias{VSA}
\title{Compute the Vowel space area from formant frequency values}
\usage{
VSA(F1, F2)
}
\arguments{
\item{F1}{A vector of F1 formant frequency measurements, one for each measured vowel.}

\item{F2}{A vector of F2 formant frequency measurements, one for each measured vowel.}
}
\value{
The vowel space area (in Hz^2). 
 
 @examples 
 data(pb)
 VSA(pb[[,"F2"]],pb[[,"F1"]])
}
\description{
This function computes the area of a vowel space by taking the 2D convex hull of the space made by 
vowels on an F1-F2 plane. The
}
