% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfd.R
\name{VSD}
\alias{VSD}
\title{Compute the Vowel space density from formant values}
\usage{
VSD(F2, F1, resolution = 0.05, grid.res = 0.01, density.threshold = 0.25)
}
\arguments{
\item{F2}{A vector of F2 formant frequency measurements, one for each measure vowel.}

\item{F1}{A vector of F1 formant frequency measurements, one for each measure vowel.}

\item{resolution}{The distance on the normalized F2-F1 space within which vowels will be counted towards the tally of vowels in close proximity for the point.}

\item{grid.res}{The spectral resolution of the analysis.}

\item{density.threshold}{The fraction of the maximum density of vowels below which the density will be considered zero.}
}
\value{
The vowel space area (in Hz^2).
}
\description{
This function computes the Vowel space density from a vector of F_1 and F_2 measurements based on the algorithm of 
\insertCitation{Story.2017.10.1121/1.4983342}{articulated}.
}
\references{
\insertAllCited{}
 
 @examples 
 data(pb)
 VSD(pb[,"F2"],pb[,"F1"]) -> ch
 #Simple but informative plot
 plot(ch,xlab="<-Back / Front -> (F2)",ylab="<-Closed / Open -> (F1)")
}
